For each column with codes add label column to data set

Adds cbs labels to the dataset that was retrieved using cbs_get_data().

cbs_add_label_columns(x, columns = colnames(x), ...)

Arguments

x

data.frame retrieved using cbs_get_data().

columns

character with the names of the columns for which labels will be added

...

not used.

Value

the original data.frame x with extra label columns. (see description)

Details

Code columns will be translated into label columns for each of the column that was supplied.

By default all code columns will be accompagnied with a label column. The name of each label column will be <code_column>_label.

See also

Examples

if (FALSE) {

# get data for main (000000) Consumer Price Index (7196ENG) for March 2000, 
 x <- cbs_get_data( id      = "7196ENG"
                  , Periods = "2000MM03"  # March 2000
                  , CPI     = "000000"    # main price index
                  )
 cbs_add_label_columns(x)
}