Gets all data from a cbs table.

Gets all data via bulk download. cbs_download_data dumps the data in (international) csv format.

cbs_download_data(
  id,
  path = file.path(id, "data.csv"),
  catalog = catalog,
  ...,
  select = NULL,
  typed = TRUE,
  verbose = FALSE,
  show_progress = interactive() && !verbose,
  base_url = getOption("cbsodataR.base_url", BASE_URL)
)

Arguments

id

of cbs open data table

path

of data file, defaults to "id/data.csv"

catalog

catalog id, can be retrieved with cbs_get_datasets()

...

optional filter statements to select rows of the data,

select

optional names of columns to be returned.

typed

Should the data automatically be converted into integer and numeric?

verbose

show the underlying downloading of the data

show_progress

show a progress bar while downloading.

base_url

optionally specify a different server. Useful for third party data services implementing the same protocol.

See also