Download a table from statistics Netherlands

This method is deprecated in favor of cbs_download_table().

download_table(id, ..., dir = id, cache = FALSE, verbose = TRUE,
  typed = FALSE, base_url = getOption("cbsodataR.base_url", BASE_URL))

Arguments

id
Identifier of CBS table (can be retrieved from cbs_get_toc())

...
Parameters passed on to cbs_download_data()

dir
Directory where table should be downloaded

cache
If metadata is cached use that, otherwise download meta data

verbose
Print extra messages what is happening.

typed
Should the data automatically be converted into integer and numeric?

base_url
optionally specify a different server. Useful for third party data services implementing the same protocol.

Value

meta data object of id cbs_get_meta().

Details

cbs_download_table retrieves all raw meta data and data and stores these as csv files in the directory specified by dir. It is possible to add a filter. A filter is specified with <column_name> = <values> in which <values> is a character vector. Rows with values that are not part of the character vector are not returned.

See also

Examples

# NOT RUN {
# download meta data and data from inflation/Consumer Price Indices
 download_table(id="7196ENG")
# }