Get meta data from table

This method is deprecated in favor of cbs_get_meta()

get_meta(
  id,
  verbose = TRUE,
  cache = FALSE,
  base_url = getOption("cbsodataR.base_url", BASE_URL)
)

Arguments

id

internal id of CBS table, can be retrieved with cbs_get_datasets()

verbose

Print extra messages what is happening.

cache

should the result be cached?

base_url

optionally specify a different server. Useful for third party data services implementing the same protocol.

Value

cbs_table object containing several data.frames with meta data (see details)

Details

The meta data of a CBS table is determined by the web api of Statistics Netherlands. cbsodataR stays close to this API. Each cbsodataR object has the following metadata items, which are all data.frames :

$TableInfos: data.frame with the descriptive publication metadata of the table, such as Title, Description, Summary etc.

$DataProperties: data.frame with the Title, Description, Unit etc. of each column in the dataset that is downloaded with cbs_get_data().

$CategoryGroups: hierarchical groupings of the code columns.

$<code column>: for each code column a data.frame with the Title, Key, Description etc. of each code / category in that column. e.g. Perioden for time codes c("2019JJ00","2018JJ00").

See also